build_acd/requirements.txt
aiohttp
jinja2
semantic_version
sh

# Testing
pytest-asyncio
asynctest
